bundled the SingleHiscoreSkillResult lookup function with the existing lookup function in HiscoreClient
