--- http-api/src/main/java/net/runelite/http/api/hiscore/SingleHiscoreSkillClient.java
@@ -1,73 +1,0 @@
-/*
- * Copyright (c) 2017. l2-
- *
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *     list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *     this list of conditions and the following disclaimer in the documentation
- *     and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package net.runelite.http.api.hiscore;
-
-import com.google.gson.JsonParseException;
-import net.runelite.http.api.RuneliteAPI;
-import okhttp3.HttpUrl;
-import okhttp3.Request;
-import okhttp3.Response;
-import okhttp3.ResponseBody;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-
-public class SingleHiscoreSkillClient
-{
-	private static final Logger logger = LoggerFactory.getLogger(SingleHiscoreSkillClient.class);
-
-	public SingleHiscoreSkillResult lookup(String username, String skill) throws IOException
-	{
-		HttpUrl.Builder builder = RuneliteAPI.getApiBase().newBuilder()
-				.addPathSegment("hiscore")
-				.addPathSegment(skill)
-				.addQueryParameter("username", username);
-
-		HttpUrl url = builder.build();
-
-		logger.debug("Built URI: {}", url);
-
-		Request request = new Request.Builder()
-				.url(url)
-				.build();
-
-		Response response = RuneliteAPI.CLIENT.newCall(request).execute();
-
-		try (ResponseBody body = response.body())
-		{
-			InputStream in = body.byteStream();
-			return RuneliteAPI.GSON.fromJson(new InputStreamReader(in), SingleHiscoreSkillResult.class);
-		}
-		catch (JsonParseException ex)
-		{
-			throw new IOException(ex);
-		}
-	}
-}
